[nfc] Extract shared Type 4 ISO-DEP NDEF logic
